refactor: model is now using modern file hierarchy
