+++ src/model.rs
@@ -1,0 +1,10 @@
+mod appstate;
+mod chat;
+mod error;
+mod hashing;
+mod log;
+mod message;
+mod pagination;
+mod refresh_token;
+mod relation;
+mod user;
