--- src/model/chat/mod.rs
@@ -1,7 +1,0 @@
-mod chat_info;
-mod chat_repository;
-mod chat;
-
-pub use chat_info::*;
-pub use chat_repository::*;
-pub use chat::*;
