--- src/model/message/mod.rs
@@ -1,7 +1,0 @@
-mod message;
-mod message_flag;
-mod message_repository;
-
-pub use message::*;
-pub use message_flag::*;
-pub use message_repository::*;
