--- src/model/misc/mod.rs
@@ -1,11 +1,0 @@
-mod appstate;
-mod error;
-mod log;
-mod pagination;
-mod hashing;
-
-pub use appstate::*;
-pub use error::*;
-pub use log::*;
-pub use pagination::*;
-pub use hashing::*;
