--- src/model/mod.rs
@@ -1,6 +1,0 @@
-pub mod user;
-pub mod chat;
-pub mod message;
-pub mod misc;
-pub mod token;
-pub mod relation;
