+++ src/model/refresh_token.rs
@@ -1,13 +1,19 @@
+mod refresh_token_flag;
+mod refresh_token_repository;
+
+pub use refresh_token_flag::*;
+pub use refresh_token_repository::*;
+
 use argon2::password_hash::rand_core::{OsRng, RngCore};
 use base64::{alphabet, engine::{general_purpose, GeneralPurpose}, Engine};
 use bson::serde_helpers::chrono_datetime_as_bson_datetime;
 use chrono::{DateTime, Duration, Utc};
+use refresh_token_flag::RefreshTokenFlag;
 use serde::Deserialize;
 use uuid::Uuid;
 
 use crate::model::user::User;
 
-use super::RefreshTokenFlag;
 
 const REFRESH_TOKEN_TTL_IN_DAYS: i64 = 365;
