+++ src/model/refresh_token/refresh_token_repository.rs
@@ -1,6 +1,6 @@
 use axum::async_trait;
 
-use crate::model::misc::ServerError;
+use crate::model::error::ServerError;
 
 use super::RefreshToken;
