--- src/model/token/mod.rs
@@ -1,7 +1,0 @@
-mod refresh_token;
-mod refresh_token_repository;
-mod refresh_token_flag;
-
-pub use refresh_token::*;
-pub use refresh_token_repository::*;
-pub use refresh_token_flag::*;
