--- src/model/user/mod.rs
@@ -1,7 +1,0 @@
-mod user;
-mod user_repository;
-mod user_flag;
-
-pub use user::*;
-pub use user_repository::*;
-pub use user_flag::*;
